+++ src/typeconvs/__init__.py
@@ -3,18 +3,19 @@ r"""Converters from :class:`str` to another type.
 These functions can be used for the ``type`` parameter in methods of
 :class:`~argparse.ArgumentParser` and :class:`~argparsebuilder.ArgParseBuilder`
 or similar use cases.
-
-All functions raise :class:`ValueError`\ s if the string cannot be converted.
 """
 
-import datetime as dt
 import locale
+import re
+from datetime import datetime, date, time
 from importlib.resources import read_text
+from operator import le, ge
 
 __version__ = read_text(__package__, 'VERSION').strip()
 
-__all__ = ['booleans', 'use_locale_default', 'boolean', 'int_conv',
-           'float_conv', 'factor_conv', 'duration', 'datetime', 'date', 'time']
+__all__ = ['booleans', 'use_locale_default', 'bool_conv', 'int_conv',
+           'float_conv', 'factor_conv', 'duration', 'datetime_conv',
+           'date_conv', 'time_conv', 'range_conv', 'sequence']
 
 use_locale_default = False
 """Default value for :func:`int_conv` and :func:`float_conv`
@@ -36,7 +37,7 @@ booleans = {
 }  #: Mapping from strings to boolean values.
 
 
-def boolean(string, values=None):
+def bool_conv(string, *, values=None):
     """Convert a string to a boolean value.
 
     The ``string`` is converted to lower case before
@@ -47,6 +48,9 @@ def boolean(string, values=None):
     :value dict: mapping from strings to booleans
     :return: converted value
     :rtype: bool
+    :raises ValueError: if the string cannot be converted
+
+    .. versionchanged:: 0.2.0 Renamed (old name: boolean)
     """
     if values is None:
         values = booleans
@@ -83,6 +87,7 @@ def int_conv(string, *, base=10, pred=None, use_locale=None):
     :type use_locale: bool or None
     :return: converted value
     :rtype: int
+    :raises ValueError: if the string cannot be converted
     """
     if (use_locale if use_locale is not None else use_locale_default):
         string = locale.delocalize(string)
@@ -103,6 +108,7 @@ def float_conv(string, *, base=10, pred=None, use_locale=None):
     :type use_locale: bool or None
     :return: converted value
     :rtype: float
+    :raises ValueError: if the string cannot be converted
     """
     if (use_locale if use_locale is not None else use_locale_default):
         string = locale.delocalize(string)
@@ -139,7 +145,7 @@ def float_conv(string, *, base=10, pred=None, use_locale=None):
     raise ValueError(f'invalid value: {string!r}')
 
 
-def factor_conv(string, *, conv=None, factors=None):
+def factor_conv(string, *, conv, factors):
     """Convert a string with a factor.
 
     The symbols from ``factors`` are compared to the end of ``string``
@@ -155,6 +161,7 @@ def factor_conv(string, *, conv=None, factors=None):
     :param conv: converter function
     :param dict factors: mapping from symbol to factor
     :return: converted value
+    :raises ValueError: if the string cannot be converted
     """
     for sym in factors:
         if string.endswith(sym):
@@ -165,7 +172,7 @@ def factor_conv(string, *, conv=None, factors=None):
     raise ValueError(f'invalid value: {string!r}')
 
 
-def datetime(string, format=None):
+def datetime_conv(string, *, format=None):
     """Convert a string to a :class:`datetime.datetime`.
 
     If ``format=None`` this function uses
@@ -177,14 +184,17 @@ def datetime(string, format=None):
     :type format: str or None
     :return: converted datetime
     :rtype: datetime.datetime
+    :raises ValueError: if the string cannot be converted
+
+    .. versionchanged:: 0.2.0 Renamed (old name: datetime)
     """
     if format is None:
-        return dt.datetime.fromisoformat(string)
+        return datetime.fromisoformat(string)
     else:
-        return dt.datetime.strptime(string, format)
+        return datetime.strptime(string, format)
 
 
-def date(string, format=None):
+def date_conv(string, *, format=None):
     """Convert a string to a class:`datetime.date`.
 
     If ``format=None`` this function uses
@@ -195,14 +205,17 @@ def date(string, format=None):
     :param format: format string
     :return: converted date
     :rtype: datetime.date
+    :raises ValueError: if the string cannot be converted
+
+    .. versionchanged:: 0.2.0 Renamed (old name: date)
     """
     if format is None:
-        return dt.date.fromisoformat(string)
+        return date.fromisoformat(string)
     else:
-        return dt.datetime.strptime(string, format).date()
+        return datetime.strptime(string, format).date()
 
 
-def time(string, format=None):
+def time_conv(string, *, format=None):
     """Convert a string to a :class:`datetime.time`.
 
     If ``format=None`` this function uses
@@ -213,14 +226,17 @@ def time(string, format=None):
     :param format: format string
     :return: converted time
     :rtype: datetime.time
+    :raises ValueError: if the string cannot be converted
+
+    .. versionchanged:: 0.2.0 Renamed (old name: time)
     """
     if format is None:
-        return dt.time.fromisoformat(string)
+        return time.fromisoformat(string)
     else:
-        return dt.datetime.strptime(string, format).time()
+        return datetime.strptime(string, format).time()
 
 
-def duration(string, use_locale=None):
+def duration(string, *, use_locale=None):
     """Convert duration string to seconds.
 
     Format: [[H:]M:]S[.f]
@@ -233,6 +249,7 @@ def duration(string, use_locale=None):
     :type use_locale: bool or None
     :return: converted duration
     :rtype: float
+    :raises ValueError: if the string cannot be converted
 
     .. versionchanged:: 0.1.1 Add parameter ``use_locale``
     """
@@ -250,3 +267,75 @@ def duration(string, use_locale=None):
     except ValueError:
         raise ValueError(f'invalid duration: {string!r}')
     return h * 3600.0 + m * 60.0 + s
+
+
+_num_re = r'(?:\d(?:\.\d*)*)+'
+_range_re = re.compile(fr'^\s*(?P<start>[-+]?{_num_re})\s*-\s*'
+                       fr'(?P<end>[-+]?{_num_re})'
+                       fr'(?:\s*/\s*(?P<step>[-+]?{_num_re}))?\s*$')
+
+
+def range_conv(string, *, conv=None):
+    """Convert a range string.
+
+    Range string: '<start>-<end>[/<step>]' (default: <step> = 1)
+
+    >>> list(range_conv('-2-4'))
+    [-2, -1, 0, 1, 2, 3, 4]
+
+    >>> list(range_conv('-2-4/2'))
+    [-2, 0, 2, 4]
+
+    >>> list(range_conv('4--2/-2'))
+    [4, 2, 0, -2]
+
+    >>> from itertools import chain
+    >>> list(chain.from_iterable(sequence('1, -2-4/2, 42', conv=range_conv)))
+    [1, -2, 0, 2, 4, 42]
+
+    :param str string: input string
+    :param conv: converter function which returns an int or a float value
+                 (default: ``int``)
+    :return: generator that yields converted values
+    :raises ValueError: if the string cannot be converted
+
+    .. versionadded:: 0.2.0
+    """
+    if not conv:
+        conv = int
+    if m := _range_re.match(string):
+        start = conv(m['start'])
+        if not isinstance(start, (int, float)):
+            raise ValueError('conv must return int or float')
+        if m['step']:
+            step = conv(m['step'])
+            if not step:
+                raise ValueError('step cannot be zero')
+        else:
+            step = 1 if isinstance(start, int) else 1.0
+        end = conv(m['end'])
+        cmp = le if step > 0 else ge
+        x = start
+        while cmp(x, end):
+            yield x
+            x += step
+    else:
+        yield conv(string)
+
+
+def sequence(string, *, conv=None, separator=','):
+    """Convert a sequence string.
+
+    >>> sequence('1, 2, 3', conv=float)
+    (1.0, 2.0, 3.0)
+
+    :param str string: input string
+    :param conv: converter function
+    :param str separator: separator between elements of sequence
+    :return: generator that yields converted values
+    :raises ValueError: if the string cannot be converted
+
+    .. versionadded:: 0.2.0
+    """
+    for s in string.split(separator):
+        yield conv(s)